requirements.txt
Pillow
pyperclip 